pyproject.toml
[tool.poetry]
name = "webpconv"
version = "0.1.0"
description = ""
authors = ["Raul.SantosSilveira-ext <Raul.SantosSilveira-ext@ab-Inbev.com >"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
PIL-Tools = "^1.1.0"
pil-lite = "^0.1.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
